Move manip/util/show_model to pydrake/visualization/model_visualizer
Move the manipulation/util/show_model.py script to
pydrake/visualization/model_visualizer.py and replace with a forwarding
script. Add additional unit tests.
