+++ bindings/pydrake/visualization/BUILD.bazel
@@ -40,6 +40,7 @@ drake_pybind_library(
     py_deps = [
         "//bindings/pydrake/multibody",
         "//bindings/pydrake/solvers",
+        "//bindings/pydrake/systems",
         "//bindings/pydrake:geometry_py",
         "//bindings/pydrake:lcm_py",
         "//bindings/pydrake:symbolic_py",
@@ -47,8 +48,10 @@ drake_pybind_library(
     ],
     py_srcs = [
         "meldis.py",
+        "model_visualizer.py",
         "plotting.py",
         "_meldis.py",
+        "_model_visualizer.py",
         "_plotting.py",
         "_visualization_extra.py",
     ],
@@ -57,6 +60,13 @@ drake_pybind_library(
 drake_py_binary(
     name = "meldis",
     srcs = ["meldis.py"],
+    visibility = ["//tools:__pkg__"],
+    deps = [":visualization"],
+)
+
+drake_py_binary(
+    name = "model_visualizer",
+    srcs = ["model_visualizer.py"],
     visibility = ["//tools:__pkg__"],
     deps = [":visualization"],
 )
@@ -90,6 +100,16 @@ drake_py_unittest(
         ":visualization",
         "//bindings/pydrake/common/test_utilities:deprecation_py",
     ],
+)
+
+drake_py_unittest(
+    name = "model_visualizer_test",
+    data = [
+        ":model_visualizer",
+        "//manipulation/util:test_models",
+        "//multibody/benchmarks/acrobot:models",
+    ],
+    deps = [":model_visualizer"],
 )
 
 drake_py_unittest(
